Refactor transactions for adding general types in redux
